functions/chunk_processor/requirements.txt
fastapi
uvicorn
google-cloud-storage
